feat: add filter query parameters
Usage: append i.e. `?type=Catalyst` to the `/all` sub route of an entity and then only the entities which's field with the key `type` is equal to `Catalyst`. Those query parameters can be chained, so if you e.g. wanted all weapons with a rarity of `3` and type `Bow`, then you would use `?type=Bow&rarity=3`. 
 
Closes #43